+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [Unreleased]
 
-* Finished the base version of the `getInterfaceID` function. Currently, it is not working properly with complex interfaces involving inheritance.
+* Finished the base version of the `getInterfaceID` function.
+* Used `Hardhat` and `solidity-ast` packages to parse the contract and get the interface ID.